Write unit tests for forgot password
